requirements.txt
streamlit>=1.30
tensorflow==2.19.0
numpy
Pillow
